Add a generic error handler
